Threw aia lib into the mix...

+++ flares/data/load.py
@@ -6,6 +6,7 @@ import re
 import shutil
 import urllib.request
 from typing import Tuple, List, Dict
+import time
 
 import astropy.coordinates
 import astropy.time
@@ -17,6 +18,9 @@ import sunpy.coordinates
 import sunpy.instr.aia
 import sunpy.map
 import sunpy.physics.differential_rotation
+from aia_lib import mov_img as aia_mov_img
+
+from PIL import Image
 
 from flares import util
 
@@ -24,10 +28,12 @@ logger = logging.getLogger(__name__)
 
 
 def sample_path(sample_id: str, output_directory: str) -> str:
-    return os.path.join(output_directory, sample_id)
+    ar_nr, p = sample_id.split("_",1)
+    return os.path.join(output_directory, ar_nr, p)
 
 
 class RequestSender(object):
+    """Downloads FITS URLs for later use in the ImageDownloader"""
     SERIES_NAMES = (
         # TODO: HMI
         "aia.lev1_vis_1h",
@@ -44,15 +50,22 @@ class RequestSender(object):
         sample_id, sample_values = sample_input
         logger.debug("Requesting data for sample %s", sample_id)
 
-        try:
-            # Perform request and provide URLs as result
-            request_urls = self._perform_request(sample_id, sample_values.start, sample_values.end)
-            self._output_queue.put((sample_id, request_urls))
-        except Exception as e:
-            logger.error("Requesting data for sample %s failed (is skipped): %s", sample_id, e)
+        retries = 15
+        while retries > 0:
+            try:
+                # Perform request and provide URLs as result
+                request_urls = self._perform_request(sample_id, sample_values.start, sample_values.end)
+            except Exception as e:
+                retries -= 1
+                logger.info("Error fetching URLs for sample %s : %s", sample_id, e)
+                time.sleep(0.5 * (15 - retries))
+            else:
+                logger.info(f'received URLs for {sample_id} after {15-retries} retries')
+                self._output_queue.put((sample_id, request_urls))
+                break
 
     def _perform_request(self, sample_id: str, start: dt.datetime, end: dt.datetime) -> List[str]:
-        client = drms.Client(email=self._notify_email)
+        client = drms.Client(email=self._notify_email, verbose=True)
         input_hours = (end - start) // dt.timedelta(hours=1)
 
         # Submit requests
@@ -91,6 +104,7 @@ class ImageLoader(object):
         logging.debug("Image loader started")
         while True:
             current_input = self._input_queue.get()
+            logging.info(f'Remaining URL sets in queue: {self._input_queue.qsize()}')
 
             # Check if done
             if current_input is None:
@@ -99,7 +113,7 @@ class ImageLoader(object):
             sample_id, records = current_input
             logger.debug("Downloading images of sample %s", sample_id)
 
-            sample_directory = os.path.join(self._output_directory, sample_id)
+            sample_directory = sample_path(sample_id, self._output_directory)
             try:
                 # Download image
                 self._download_images(sample_directory, records)
@@ -111,11 +125,14 @@ class ImageLoader(object):
                 logger.error("Error while downloading data for sample %s (is skipped): %s", sample_id, e)
 
                 # Delete sample directory because it contains inconsistent data
-                shutil.rmtree(sample_directory, ignore_errors=True)
+                # TODO: Just delete this 1 input data, not the entire fits folder...
+                #shutil.rmtree(sample_directory, ignore_errors=True)
 
     def _download_images(self, sample_directory: str, records: List[Tuple[str, str]]):
         fits_directory = os.path.join(sample_directory, "_fits_temp")
         os.makedirs(fits_directory)
+
+        logger.info(f'Downloading {len(records)} FITS files into {fits_directory}...')
 
         for record, url in records:
             # TODO: This does not work with HMI
@@ -128,9 +145,21 @@ class ImageLoader(object):
             record_date = dt.datetime.strptime(record_date_raw, self.RECORD_DATE_FORMAT)
 
             output_file_name = f"{record_date:%Y-%m-%dT%H%M%S}_{record_wavelength}.fits"
-            urllib.request.urlretrieve(url, os.path.join(fits_directory, output_file_name))
+            fp = os.path.join(fits_directory, output_file_name)
+            if not os.path.isfile(fp): #TODO: Check for corruption, incomplete files
+                retries = 15
+                while retries > 0:
+                    try:
+                        urllib.request.urlretrieve(url, os.path.join(fits_directory, output_file_name))
+                    except Exception as e:
+                        retries -= 1
+                        logger.info("Error fetching FITS %s : %s", url, e)
+                        time.sleep(0.5 * (15 - retries))
+                    else:
+                        logger.info(f'{15-retries} retries')
+                        break
 
-        logger.debug("Downloaded %d files to %s", len(records), fits_directory)
+        logger.info("Downloaded %d files to %s", len(records), fits_directory)
 
 
 class OutputProcessor(object):
@@ -167,6 +196,7 @@ class OutputProcessor(object):
         logging.debug("Output processor started")
         while True:
             sample_id = self._input_queue.get()
+            logging.info(f'Remaining URL sets in queue: {self._input_queue.qsize()}')
 
             # Check if done
             if sample_id is None:
@@ -174,7 +204,7 @@ class OutputProcessor(object):
 
             logger.debug("Processing sample %s", sample_id)
 
-            sample_directory = os.path.join(self._output_directory, sample_id)
+            sample_directory = sample_path(sample_id, self._output_directory)
             fits_directory = os.path.join(sample_directory, "_fits_temp")
 
             try:
@@ -184,10 +214,11 @@ class OutputProcessor(object):
                 logger.error("Error while processing data for sample %s (is skipped): %s", sample_id, e)
 
                 # Delete sample directory because it contains inconsistent data
-                shutil.rmtree(sample_directory, ignore_errors=True)
+                #TODO:shutil.rmtree(sample_directory, ignore_errors=True)
             finally:
                 # Delete fits directory in any case to avoid space issues
-                shutil.rmtree(fits_directory, ignore_errors=True)
+                #TODO:shutil.rmtree(fits_directory, ignore_errors=True)
+                print(f'Would have deleted FITS files {fits_directory}')
 
     def _process_output(self, sample_id: str, input_directory: str, output_directory: str):
 
@@ -240,9 +271,14 @@ class OutputProcessor(object):
                     logger.warning("Discarding wavelength %s for sample %s", current_wavelength, sample_id)
                     continue
 
+                # create the image
+                current_img = aia_mov_img.process_img(current_map)
+                assert current_map.data.shape[0] == current_img.shape[0]
+                assert current_map.data.shape[1] == current_img.shape[1]
+
                 # Convert to level 1.5
-                assert current_map.processing_level == 1.0
-                current_map = sunpy.instr.aia.aiaprep(current_map)
+                #assert current_map.processing_level == 1.0
+                #current_map = sunpy.instr.aia.aiaprep(current_map)
 
                 # Find coordinates of closest active region event which started before the image
                 image_time = current_map.date
@@ -273,7 +309,7 @@ class OutputProcessor(object):
                 # Sunpy maps assume a carthesian coordinate system which is already incorporated in the pixel conversion
                 patch_start_y = center_y - self.OUTPUT_SHAPE[0] // 2
                 patch_start_x = center_x - self.OUTPUT_SHAPE[1] // 2
-                current_patch = current_map.data[
+                current_patch = current_img[
                     patch_start_y:patch_start_y + self.OUTPUT_SHAPE[0],
                     patch_start_x:patch_start_x + self.OUTPUT_SHAPE[1],
                 ]
@@ -282,13 +318,24 @@ class OutputProcessor(object):
                 # TODO: Convert back to int16? (results in way better compression)
 
                 # Save patch
-                output_arrays[current_wavelength] = current_patch
+                #output_arrays[current_wavelength] = current_patch
+
+                # what happens when converting to other formats?
+                patch_16bitint = (np.round(current_patch)).astype(np.int16)
+                #logger.info(f'Total diff int16={np.sum(np.abs(patch_16bitint - current_patch))}')
+                #logger.info(f'Max diff int16={np.amax(np.abs(patch_16bitint - current_patch))}')
+
+                # Save as image
+                output_file_path = os.path.join(output_directory, current_datetime.strftime("%Y-%m-%dT%H%M%S") + "__" + str(current_wavelength))
+                im = Image.fromarray(patch_16bitint)
+                im.save(output_file_path, "PNG")
+
 
             # Save patches as compressed numpy file
-            output_file_path = os.path.join(output_directory, current_datetime.strftime("%Y-%m-%dT%H%M%S"))
-            np.savez_compressed(output_file_path, **output_arrays)
+            #output_file_path = os.path.join(output_directory, current_datetime.strftime("%Y-%m-%dT%H%M%S"))
+            #np.savez_compressed(output_file_path, **output_arrays)
 
-        logger.debug("Created sample %s output", sample_id)
+        logger.info("Created sample %s output", sample_id)
 
     @classmethod
     def _is_usable(cls, target: sunpy.map.sources.AIAMap) -> bool:
